refactor(select-game): extract components and hooks for better maintainability
Move ModeCard and LevelCard to separate components
Create useSelectGame hook to centralize game selection logic
Add types.ts for shared interfaces
Clean up selectGame.tsx by removing embedded components
